requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client>=3.0.0
langchain-pinecone
langchain_community
langchain_experimental
langchain-nomic
langchain-google-genai
-e .